Time: 28 ms (55.48%) | Memory: 28.9 MB (90.32%) - LeetSync

+++ 3496-minimum-number-of-seconds-to-make-mountain-height-zero/minimum-number-of-seconds-to-make-mountain-height-zero.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     long long minNumberOfSeconds(int mountainHeight, vector<int>& workerTimes) {
        long long H=*max_element(workerTimes.begin(), workerTimes.end());
-       long long start=1,end=static_cast<long long>(H)*mountainHeight*(mountainHeight+1)/2;
+       long long start=1,end=(H)*mountainHeight*(mountainHeight+1)/2;
        long long ans=0;
        while(start<=end)
        {
